# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Website Builder Platform API — bilingual (AR/EN)
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tokens/usage:
    get:
      operationId: getTokenUsage
      tags: [tokens]
      summary: Token usage
      description: Returns token usage filtered by date range and project
      parameters:
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: startDate
          in: query
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          schema:
            type: string
            format: date
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, project, agent]
      responses:
        "200":
          description: Usage data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenUsageResponse"
  /tokens/limits:
    get:
      operationId: getTokenLimits
      tags: [tokens]
      summary: Get limits
      description: Returns the user's current spending limits
      responses:
        "200":
          description: Current limits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenLimits"
    patch:
      operationId: updateTokenLimits
      tags: [tokens]
      summary: Update limits
      description: Update daily or monthly spending limits
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTokenLimitsRequest"
      responses:
        "200":
          description: Updated limits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenLimits"
...
  /tokens/notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [tokens]
      summary: Mark notification read
      description: Marks a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
...
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
...
    TokenUsageEntry:
      type: object
      properties:
        date:
          type: string
          format: date
        projectId:
          type: string
          format: uuid
        projectName:
          type: string
        agentType:
          type: string
        tokensInput:
          type: integer
        tokensOutput:
          type: integer
        costUsd:
          type: number
      required:
        - tokensInput
        - tokensOutput
        - costUsd
    TokenUsageResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TokenUsageEntry"
        totalTokens:
          type: integer
        totalCostUsd:
          type: number
      required:
        - data
        - totalTokens
        - totalCostUsd
    TokenLimits:
      type: object
      properties:
        dailyLimitUsd:
          type: number
        monthlyLimitUsd:
          type: number
        perProjectLimitUsd:
          type: number
        dailyUsedUsd:
          type: number
        monthlyUsedUsd:
          type: number
      required:
        - dailyLimitUsd
        - monthlyLimitUsd
        - dailyUsedUsd
        - monthlyUsedUsd
    UpdateTokenLimitsRequest:
      type: object
      properties:
        dailyLimitUsd:
          type: number
        monthlyLimitUsd:
          type: number
        perProjectLimitUsd:
          type: number
...